+++ src/pas/plugins/authomatic/plugin.py
@@ -47,6 +47,7 @@ manage_addAuthomaticPluginForm = PageTemplateFile(
     pas_interfaces.IAuthenticationPlugin,
     pas_interfaces.IPropertiesPlugin,
     pas_interfaces.IUserEnumerationPlugin,
+    pas_interfaces.IRolesPlugin
 )
 class AuthomaticPlugin(BasePlugin):
     """Authomatic PAS Plugin
@@ -286,6 +287,24 @@ class AuthomaticPlugin(BasePlugin):
         if sort_by in ['id', 'login']:
             return sorted(ret, key=itemgetter(sort_by))
         return ret
+
+    @security.private
+    def getRolesForPrincipal(self, user, request=None):
+        """ Fullfill RolesPlugin requirements """
+        identity = self._useridentities_by_userid.get(user.getId(), None)
+        if not identity:
+            return ()
+        keys = [key for key in identity._identities.keys()]
+        provider_id = keys[0]
+
+        if 'roles' in identity._identities[provider_id].keys():
+            roles = identity._identities[provider_id]['roles']
+            if isinstance(roles, list):
+                return tuple(roles)
+            else:
+                return ()
+        else:
+            return ()
 
 
 InitializeClass(AuthomaticPlugin)